+++ src/python/pants/backend/experimental/go/register.py
@@ -3,7 +3,7 @@
 
 from pants.backend.go import build, module, pkg, tailor, target_type_rules
 from pants.backend.go import target_types as go_target_types
-from pants.backend.go.goals import custom_goals
+from pants.backend.go.goals import custom_goals, package_binary
 from pants.backend.go.subsystems import golang
 from pants.backend.go.target_types import GoBinary, GoExternalPackageTarget, GoModule, GoPackage
 from pants.backend.go.util_rules import import_analysis, sdk
@@ -25,4 +25,5 @@ def rules():
         *tailor.rules(),
         *target_type_rules.rules(),
         *custom_goals.rules(),
+        *package_binary.rules(),
     ]
