+++ src/python/pants/backend/go/build.py
@@ -17,20 +17,13 @@ from pants.backend.go.pkg import (
     is_third_party_package_target,
 )
 from pants.backend.go.target_types import (
-    GoBinaryMainAddress,
     GoExternalModulePathField,
     GoExternalModuleVersionField,
     GoPackageSources,
 )
 from pants.backend.go.util_rules.import_analysis import ResolvedImportPathsForGoLangDistribution
 from pants.backend.go.util_rules.sdk import GoSdkProcess
-from pants.build_graph.address import Address, AddressInput
-from pants.core.goals.package import (
-    BuiltPackage,
-    BuiltPackageArtifact,
-    OutputPathField,
-    PackageFieldSet,
-)
+from pants.build_graph.address import Address
 from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
 from pants.engine.addresses import Addresses
 from pants.engine.engine_aware import EngineAwareParameter
@@ -38,15 +31,7 @@ from pants.engine.fs import AddPrefix, CreateDigest, Digest, FileContent, MergeD
 from pants.engine.internals.selectors import Get, MultiGet
 from pants.engine.process import ProcessResult
 from pants.engine.rules import collect_rules, rule
-from pants.engine.target import (
-    Dependencies,
-    DependenciesRequest,
-    TransitiveTargets,
-    TransitiveTargetsRequest,
-    UnexpandedTargets,
-    WrappedTarget,
-)
-from pants.engine.unions import UnionRule
+from pants.engine.target import Dependencies, DependenciesRequest, UnexpandedTargets, WrappedTarget
 from pants.util.ordered_set import FrozenOrderedSet, OrderedSet
 from pants.util.strutil import pluralize
 
@@ -66,14 +51,6 @@ class BuildGoPackageRequest(EngineAwareParameter):
 class BuiltGoPackage:
     import_path: str
     object_digest: Digest
-
-
-@dataclass(frozen=True)
-class GoBinaryFieldSet(PackageFieldSet):
-    required_fields = (GoBinaryMainAddress,)
-
-    main_address: GoBinaryMainAddress
-    output_path: OutputPathField
 
 
 @dataclass(frozen=True)
@@ -308,78 +285,5 @@ async def build_target(
     return BuiltGoPackage(import_path=import_path, object_digest=output_digest)
 
 
-@rule
-async def package_go_binary(
-    field_set: GoBinaryFieldSet,
-) -> BuiltPackage:
-    main_address = field_set.main_address.value or ""
-    main_go_package_address = await Get(
-        Address,
-        AddressInput,
-        AddressInput.parse(main_address, relative_to=field_set.address.spec_path),
-    )
-    wrapped_main_go_package_target = await Get(WrappedTarget, Address, main_go_package_address)
-    main_go_package_target = wrapped_main_go_package_target.target
-    built_main_go_package = await Get(
-        BuiltGoPackage, BuildGoPackageRequest(address=main_go_package_target.address, is_main=True)
-    )
-
-    transitive_targets = await Get(
-        TransitiveTargets, TransitiveTargetsRequest(roots=[main_go_package_target.address])
-    )
-    buildable_deps = [
-        tgt
-        for tgt in transitive_targets.dependencies
-        if is_first_party_package_target(tgt) or is_third_party_package_target(tgt)
-    ]
-
-    built_transitive_go_deps_requests = [
-        Get(BuiltGoPackage, BuildGoPackageRequest(address=tgt.address)) for tgt in buildable_deps
-    ]
-    built_transitive_go_deps = await MultiGet(built_transitive_go_deps_requests)
-
-    gathered_imports = await Get(
-        GatheredImports,
-        GatherImportsRequest(
-            packages=FrozenOrderedSet(built_transitive_go_deps),
-            include_stdlib=True,
-        ),
-    )
-
-    input_digest = await Get(
-        Digest, MergeDigests([gathered_imports.digest, built_main_go_package.object_digest])
-    )
-
-    output_filename = PurePath(field_set.output_path.value_or_default(file_ending=None))
-    result = await Get(
-        ProcessResult,
-        GoSdkProcess(
-            input_digest=input_digest,
-            command=(
-                "tool",
-                "link",
-                "-importcfg",
-                "./importcfg",
-                "-o",
-                f"./{output_filename.name}",
-                "-buildmode=exe",  # seen in `go build -x` output
-                "./__pkg__.a",
-            ),
-            description="Link Go binary.",
-            output_files=(f"./{output_filename.name}",),
-        ),
-    )
-
-    renamed_output_digest = await Get(
-        Digest, AddPrefix(result.output_digest, str(output_filename.parent))
-    )
-
-    artifact = BuiltPackageArtifact(relpath=str(output_filename))
-    return BuiltPackage(digest=renamed_output_digest, artifacts=(artifact,))
-
-
 def rules():
-    return [
-        *collect_rules(),
-        UnionRule(PackageFieldSet, GoBinaryFieldSet),
-    ]
+    return collect_rules()
